requirements.txt
Click>=6.0
requests
pytz
suds-jurko>=0.6
future